requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
sentence-transformers>=2.2.2
faiss-cpu>=1.9.0
python-dotenv>=1.0.0
pydantic>=2.10.0
numpy>=1.26.0
requests>=2.31.0
openai>=1.54.0
anthropic>=0.39.0
google-genai>=0.3.0